# spec/openapi.yaml
openapi: 3.0.0
info:
  title: Bicpid API
  description: API for managing PIDs across namespaces and user accounts with API keys.
  version: 3.0.0
  license:
    name: Creative Commons Attribution-ShareAlike 4.0 International
    url: https://creativecommons.org/licenses/by-sa/4.0/
...
servers:
  - url: /api/v2
paths:
  /resolver:
    get:
      operationId: getResolverInfo
      summary: Return information regarding the resolvers' limits and other configuration.
      tags:
        - resolver
      responses:
        '200':
          description: Limits successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
        '404':
          description: Configuration is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfoUnavailable'
  /resolver/namespaces:
    get:
      operationId: listNamespaces
      summary: List namespaces available on the resolver.
      tags:
        - namespaces

      parameters:
        - name: tag
          description: |
            Filter namespaces by their tag.
            If omitted, namespaces of all tags are included.
            If provided with an empty value, the value is treated as the empty string.
          in: query
          required: false
          schema:
            type: string
          allowEmptyValue: true
        - name: limit
          in: query
          required: false
          description: |
            Maximum number of items to return.
            If omitted, a default value is applied.
            The maximum possible value may be clamped, and values larger than it may be silently truncated.
            If the value is a negative integer, a '400' response may be returned.
          schema:
            type: integer
            minimum: 0
        - name: offset
          in: query
          required: false
          description: |
            Number of items to skip before starting to collect the result set.
            If omitted, the value is treated as 0.
            If the value is negative or not an integer, a '400' response may be returned.
          schema:
            type: integer
            minimum: 0

      responses:
        '200':
          description: |
            A paginated list of namespaces, sorted by their id in increasing order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNamespacesResponse'
        '400':
          description: Invalid query parameter (e.g. limit or offset).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorInvalidQueryParameter'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'

    post:
      operationId: createNamespace
      summary: Creates a new namespace.
      description: |
        Creates a new namespace with a server-assigned id.
        When authentication is required, the authenticated user becomes the namespace owner
        and receives manager permissions for that namespace automatically.
      tags:
        - namespaces

      requestBody:
        required: true
        description: |
          A namespace to create.
          A limit on the maximum request body size MAY be enforced.
          If such a limit is exceeded, a '413' response is returned.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamespaceCreateRequest'

      responses:
        '201':
          description: Namespace created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
        '400':
          description: Invalid request body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodyMissing'
                  - $ref: '#/components/schemas/ErrorBodyInvalidJSON'
        '413':
          description: Request body exceeds the configured maximum size.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodySizeExceeded'
        '500':
          description: Internal server error or ID generation failure.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'
                  - $ref: '#/components/schemas/ErrorBadIDGeneration'
        '503':
          description: Could not allocate a namespace id after repeated attempts.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorInsufficientEntropy'
  /resolver/resources/count:
    get:
      operationId: countAllResources
      summary: Return the total number of resources (issued PIDs) stored across all namespaces.
      tags:
        - resources
      responses:
        '200':
          description: |
            Total count of resources in the resolver.
            Includes soft-deleted resources.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceCountResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'
  /resolver/namespaces/{namespace}:
    get:
      operationId: getNamespace
      summary: Retrieve a single namespace by id.
      tags:
        - namespaces
      parameters:
        - name: namespace
          in: path
          required: true
          description: |
            Server-assigned namespace id returned from create namespace.
          schema:
            $ref: '#/components/schemas/NamespaceId'
      responses:
        '200':
          description: Namespace found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
        '400':
          description: Invalid namespace id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInvalidNamespaceID'
        '404':
          description: Namespace does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNamespaceNotFound'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'
...
  /resolver/namespaces/{namespace}/resources:
    get:
      operationId: listResources
      summary: List resources within a namespace.
      tags:
        - resources

      parameters:
        - name: namespace
          in: path
          required: true
          description: |
            Server-assigned namespace id returned from create namespace.
          schema:
            $ref: '#/components/schemas/NamespaceId'
        - name: tag
          description: |
            Filter resources by tag.
            If omitted, resources of all tags are included.
            If provided with an empty value, the value is treated as the empty string.
          in: query
          required: false
          schema:
            type: string
          allowEmptyValue: true
        - name: deleted
          description: |
            Filter resources by their deletion status.
            If true, return only deleted resources.
            If false, return only non-deleted resources.
            If omitted, resources of both deletion states are included.
          in: query
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          description: |
            Maximum number of items to return.
            If omitted, a default value is applied.
            The maximum possible value MAY be clamped.
            Negative values or non-integers yield a '400' response.
          schema:
            type: integer
            minimum: 0
        - name: offset
          in: query
          required: false
          description: |
            Number of items to skip before starting to collect the result set.
            If omitted, the value is treated as 0.
            If the value is negative or not an integer, a '400' response may be returned.
          schema:
            type: integer
            minimum: 0

      responses:
        '200':
          description: A paginated list of resources within the namespace, ordered by their pid in increasing order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResourcesResponse'
        '400':
          description: Invalid namespace id or query parameter.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorInvalidNamespaceID'
                  - $ref: '#/components/schemas/ErrorInvalidQueryParameter'
        '404':
          description: Namespace does not exist.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorNamespaceNotFound'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'

    post:
      operationId: createResource
      summary: Create a resource within a namespace.
      description: |
        A new PID is generated, unique within the namespace, and returned with the response.
      tags:
        - resources

      parameters:
        - name: namespace
          in: path
          required: true
          description: |
            Server-assigned namespace id returned from create namespace.
          schema:
            $ref: '#/components/schemas/NamespaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreateRequest'

      responses:
        '201':
          description: Resource created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '400':
          description: Invalid request body, JSON, or namespace id in the path.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodyMissing'
                  - $ref: '#/components/schemas/ErrorBodyInvalidJSON'
                  - $ref: '#/components/schemas/ErrorInvalidNamespaceID'
        '413':
          description: Request body exceeds the configured maximum size.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodySizeExceeded'
        '404':
          description: Namespace does not exist.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorNamespaceNotFound'
        '500':
          description: Internal server error or PID generation failure.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'
                  - $ref: '#/components/schemas/ErrorBadIDGeneration'
        '503':
          description: Could not allocate a PID after repeated attempts.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorInsufficientEntropy'
  /resolver/namespaces/{namespace}/resources:batch:
    post:
      operationId: batchCreateResources
      summary: |
        Create several resources within a namespace.
        This is equivalent to calling the 'createResource' operation repeatedly.
        If a single resource fails to be created, the entire batch operation fails.
      tags:
        - resources

      parameters:
        - name: namespace
          in: path
          required: true
          description: |
            Server-assigned namespace id.
          schema:
            $ref: '#/components/schemas/NamespaceId'
      requestBody:
        required: true
        description: |
          A list of resources to create.
          Limits on how many resources can be created in a single request MAY be enforced.
          Such limits can take the form of a maximum number of resources, a maximum body size, or both.
          If the maximum number of resources is exceeded, a '422' response is returned.
          If the maximum body size is exceeded, a '413' response is returned.
        content:
          application/json:
            schema:
              type: array

              items:
                $ref: '#/components/schemas/ResourceCreateRequest'

      responses:
        '201':
          description: Resources created successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceResponse'
        '400':
          description: Invalid request body, JSON, or namespace id in the path.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodyMissing'
                  - $ref: '#/components/schemas/ErrorBodyInvalidJSON'
                  - $ref: '#/components/schemas/ErrorInvalidNamespaceID'
        '413':
          description: Request body exceeds the configured maximum size.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodySizeExceeded'
        '404':
          description: Namespace does not exist.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorNamespaceNotFound'
        '422':
          description: Number of items in the batch exceeds the configured maximum.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorItemLimitExceeded'
        '500':
          description: Internal server error or PID generation failure.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'
                  - $ref: '#/components/schemas/ErrorBadIDGeneration'
        '503':
          description: Could not allocate PIDs after repeated attempts.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorInsufficientEntropy'
  /resolver/namespaces/{namespace}/resources/{pid}:
    get:
      operationId: getResource
      summary: |
        Retrieve a resource by PID within a namespace.
      tags:
        - resources

      parameters:
        - name: namespace 
          in: path
          required: true
          description: |
            Server-assigned namespace id.
          schema:
            $ref: '#/components/schemas/NamespaceId'
        - name: pid
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/PID'

      responses:
        '200':
          description: Resource retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '400':
          description: Invalid namespace id or PID in the path.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorInvalidNamespaceID'
                  - $ref: '#/components/schemas/ErrorInvalidPID'
        '404':
          description: Namespace or resource does not exist.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorNamespaceNotFound'
                  - $ref: '#/components/schemas/ErrorResourceNotFound'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'

    patch:
      operationId: updateResource
      summary: |
        Updates a specific resource identified by a PID in a namespace.
      tags:
        - resources

      parameters:
        - name: namespace
          in: path
          required: true
          description: |
            Server-assigned namespace id.
          schema:
            $ref: '#/components/schemas/NamespaceId'
        - name: pid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: |
          The new contents of the resource to update.
          The resource update request uses patch semantics: If a field is omitted, it is not be changed.
          Regardless of the presence or absence of a field, all fields will always be returned in the response.

          Note that empty updates are permitted, and will simply return the current resource contents as is (provided they exist).

          A limit on the maximum request body size may be enforced.
          If such a limit is exceeded, a '413' response is returned.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceUpdateRequest'
      responses:
        '200':
          description: Resource updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '400':
          description: Invalid request body, JSON, namespace id, or PID.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodyMissing'
                  - $ref: '#/components/schemas/ErrorBodyInvalidJSON'
                  - $ref: '#/components/schemas/ErrorInvalidNamespaceID'
                  - $ref: '#/components/schemas/ErrorInvalidPID'
        '404':
          description: Namespace or resource does not exist.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorNamespaceNotFound'
                  - $ref: '#/components/schemas/ErrorResourceNotFound'
        '413':
          description: Request body exceeds the configured maximum size.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodySizeExceeded'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'
  /user/:
    get:
      operationId: getCurrentUser
      summary: Return information about the currently authenticated user.
      tags:
        - user
      responses:
        '200':
          description: User information successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Authentication is required or the bearer token is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'

    patch:
      operationId: updateCurrentUser
      summary: Update the currently authenticated user account.
      tags:
        - user
      requestBody:
        required: true
        description: |
          Fields to update on the user account.
          A limit on the maximum request body size MAY be enforced.
          If such a limit is exceeded, a '413' response is returned.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User account updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Invalid request body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodyMissing'
                  - $ref: '#/components/schemas/ErrorBodyInvalidJSON'
        '401':
          description: Authentication is required or the bearer token is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: The authenticated user is not allowed to perform this update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '413':
          description: Request body exceeds the configured maximum size.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodySizeExceeded'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'

    post:
      operationId: createUser
      summary: Create a new user account.
      tags:
        - user
      requestBody:
        required: true
        description: |
          A user account to create.
          A limit on the maximum request body size MAY be enforced.
          If such a limit is exceeded, a '413' response is returned.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
        '400':
          description: Invalid request body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodyMissing'
                  - $ref: '#/components/schemas/ErrorBodyInvalidJSON'
        '401':
          description: Authentication is required or the bearer token is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: The authenticated user is not allowed to create user accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
        '409':
          description: The username is already in use.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDuplicateUsername'
        '413':
          description: Request body exceeds the configured maximum size.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorBodySizeExceeded'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDatabaseError'
...
components:
  schemas:
...
    ErrorDatabaseError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: string
              enum:
                - database_error
          description: |
            An unspecified internal problem occurred while interacting with the database.
            This may be due to the database being offline, or due to data corruption. 
...
    ErrorInvalidQueryParameter:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: string
              enum:
                - invalid_query_parameter
          description: A query parameter was invalid.
...
    ErrorInvalidPID:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: string
              enum:
                - invalid_pid
          description: The PID was invalid.
...
    ErrorInfoUnavailable:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: string
              enum:
                - info_unavailable
          description: |
            Resolver configuration info is unavailable.
            The resolver may choose to hide this information for security reasons.
...
    InfoResponse:
      description:
        Public resolver limits and configuration.
        A zero value indicates that no limit is enforced.
      type: object
      required:
        - max_body_bytes
        - default_page_limit
        - max_page_limit
        - max_batch_items
      properties:
        max_body_bytes:
          description: |
            Maximum permitted size of the request body in bytes.
          type: integer
          format: int64
          minimum: 0
        default_page_limit:
          description: |
            Maximum number of items per page.
          type: integer
          format: int64
          minimum: 0
        max_page_limit:
          description: |
            Maximum number of items per page in a response.
          type: integer
          format: int64
          minimum: 0
        max_batch_items:
          description: |
            Maximum number of items in a batch.
          type: integer
          format: int64
          minimum: 0
...
    PaginatedNamespacesResponse:
      description: |
        A single page of namespace responses.
        The 'total' field is the total number of namespaces after applying the tag filter (if any), ignoring 'limit' and 'offset'.
        The 'offset' field is the effective offset used to produce this response.
      type: object
      required:
        - total
        - offset
        - items
      properties:
        total:
          type: integer
          minimum: 0
          example: 1
          description: Total number of items available.
        offset:
          type: integer
          minimum: 0
          example: 0
          description: Offset of the current response.
        items:
          type: array
          items:
            $ref: '#/components/schemas/Namespace'
...